requirements.txt
python-dotenv

# LangChain core packages
langchain
langchain-core
langchain-text-splitters
langchain-community

# LangSmith tracing and observability
langsmith

# OpenAI integration (embeddings + ChatOpenAI)
langchain-openai
openai

# Grok/xAI (OpenAI-compatible — optional)
# If Grok ends up requiring a dedicated package later, add it here.
# For now we rely on OpenAI-compatible client.


# Gemini / Google Generative AI
langchain-google-genai
google-generativeai

# Anthropic / Claude
langchain-anthropic
anthropic

# PDF loader
pypdf


# Vector DB
faiss-cpu

# Graph processing
networkx

# BM25 Retrieve
rank_bm25

# App UI
streamlit
watchdog
